[TAS-139] Design: 친구 요청이 많을 경우 overflow 발생 해결 & 폴더 정리
